Restarted Project
I have started a complete redesign for the project, I had lost vision for this version of code thus a redesign was needed.

I have now created a more structured plan for working through the project and will begin working from the begging soon.

The first step is to implement the complete game mechanics for my Cut the Rope clone.