Fixed Linker Issue
Fixed an issue from an unkown source. The program was unable to link the b2World, but after transfering over a different (but seemingly the exact same) and now the program compiles.

The next step is to implement simple physics into the game and prepare for the main gameplay mechanics.