Bouncing Ball and angled platform
I have implemented a basic level with a ball which falls and bounces off the angled platfom. The ball will bounce around constrained to the level before coming to a complete stop.

The next step is to add in a rope to the game which will suspend the ball in the air.