Rope Implementation
I have created a basic rope building runction which will allow for someone to create a rope from one object down.

Next I need to work out the issue with drawing the level bounds/platforms.

Once that is completed I will then suspend a platform from the rope and attempt to bounce the ball off of it.

I will also need to perfom some cleanup and management of the class to make it work better, for example adding in specified anchor points for the initial and final bodies the rope connects to.