First Level.
I have almost finished creating the hardcoded base level. The Next step is to add the Ball Class into the Level class allowing for it to be easily attached and suspended fron the top of the level.

After I will work on some tweeks to the level size as well as doing some general clean-up making the game code far easier to read.