Mouse Tool
I have added a class which will hold all of the mouse's functionality. It currently allows for a player to draw a line from one point to another by pressing down the LMB and relesing it at the destination.

The next step is to add the functionality where it will cut ropes it overlaps.