Game Finshed
I have finished the game. the only issue is the text within the game, i belive that this is due to the scale of the game. either way I have no idea how to fix this.

The next step is to do some testing and begin to write the repot.

I also need to do some general cleanup for the game as well.